pyproject.toml
[tool.poetry]
name = "reservation"
version = "0.1.0"
description = ""
authors = ["jslee <dlwhdtjd098@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = ">=4.1,<5.0"
djangorestframework = "^3.15.1"
markdown = "^3.6"
django-filter = ">=23.5,<24.0"
psycopg2-binary = "^2.9.9"
djangorestframework-simplejwt = "^5.3.1"
pytz = "^2024.1"
drf-spectacular = "^0.27.2"
django-redis = "^5.4.0"
redis = "^5.0.7"
pandas = "^2.2.2"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.0.0"
pyright = "^1.1.358"
ruff = "^0.3.7"
reorder-python-imports = "^3.12.0"
pre-commit = "^3.7.0"
black = "^24.4.0"
typing-extensions = "^4.12.2"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.3.0"


[tool.poetry.group.test.dependencies]
pytest = "^8.1.1"
pytest-cov = "^5.0.0"
pytest-django = "^4.8.0"
factory-boy = "^3.3.0"
pytest-xdist = "^3.5.0"
pytest-env = "^1.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
